Unify the formula wrappers behind a marker type
After the port to the unified component-graph Formula, AggregationFormula
and CoalesceFormula had byte-identical bodies (field, Display,
FormulaSubscriber, construction); only their GraphFormulaProvider impls
differed. Collapse them into one GraphFormula<M, K> keyed by an
Aggregation/Coalesce marker, keep the two provider impls, and alias the
old names. The FormulaParams argument bundle and its From conversions
are replaced by a plain GraphFormula::new.
